refactor(eventrecorder): move configuration to eventrecorder package
Avoid import cycles when moving notifier configuration types from config
to their own package in the future.
