+++ eventrecorder/config.go
@@ -1,0 +1,79 @@
+// Copyright The Prometheus Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package eventrecorder provides a structured event recorder for
+// significant Alertmanager events.  Events are serialized as JSON and
+// fanned out to one or more configured destinations (JSONL file,
+// webhook, etc.).
+//
+// RecordEvent never blocks the caller: events are serialized and
+// placed on a bounded in-memory queue.  A background goroutine
+// drains the queue and sends to destinations.  If the queue is full,
+// events are dropped and a metric is incremented.
+package eventrecorder
+
+import (
+	"errors"
+	"fmt"
+
+	commoncfg "github.com/prometheus/common/config"
+	"github.com/prometheus/common/model"
+
+	amcommoncfg "github.com/prometheus/alertmanager/config/common"
+)
+
+// EventRecorderConfig configures the event recorder feature.
+type EventRecorderConfig struct {
+	Outputs []EventRecorderOutput `yaml:"outputs,omitempty" json:"outputs,omitempty"`
+}
+
+// EventRecorderOutput configures a single event recorder output destination.
+type EventRecorderOutput struct {
+	Type       string                      `yaml:"type" json:"type"`
+	Path       string                      `yaml:"path,omitempty" json:"path,omitempty"`
+	URL        *amcommoncfg.SecretURL      `yaml:"url,omitempty" json:"url,omitempty"`
+	HTTPConfig *commoncfg.HTTPClientConfig `yaml:"http_config,omitempty" json:"http_config,omitempty"`
+	// Timeout for webhook HTTP requests (default 10s).
+	Timeout model.Duration `yaml:"timeout,omitempty" json:"timeout,omitempty"`
+	// Workers is the number of concurrent webhook delivery goroutines
+	// (default 4).  Only applicable to webhook outputs.
+	Workers int `yaml:"workers,omitempty" json:"workers,omitempty"`
+	// MaxRetries is the maximum number of delivery attempts per event
+	// (default 3).  Only applicable to webhook outputs.
+	MaxRetries int `yaml:"max_retries,omitempty" json:"max_retries,omitempty"`
+	// RetryBackoff is the base backoff duration between retry attempts
+	// (default 500ms).  Successive attempts use exponential backoff
+	// (base * 2^attempt).  Only applicable to webhook outputs.
+	RetryBackoff model.Duration `yaml:"retry_backoff,omitempty" json:"retry_backoff,omitempty"`
+}
+
+// UnmarshalYAML implements the yaml.Unmarshaler interface for EventRecorderOutput.
+func (o *EventRecorderOutput) UnmarshalYAML(unmarshal func(any) error) error {
+	type plain EventRecorderOutput
+	if err := unmarshal((*plain)(o)); err != nil {
+		return err
+	}
+	switch o.Type {
+	case "file":
+		if o.Path == "" {
+			return errors.New("event_recorder file output requires a path")
+		}
+	case "webhook":
+		if o.URL == nil {
+			return errors.New("event_recorder webhook output requires a url")
+		}
+	default:
+		return fmt.Errorf("unknown event_recorder output type %q, must be \"file\" or \"webhook\"", o.Type)
+	}
+	return nil
+}
