+++ eventrecorder/webhook_test.go
@@ -28,7 +28,6 @@ import (
 	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/require"
 
-	"github.com/prometheus/alertmanager/config"
 	amcommoncfg "github.com/prometheus/alertmanager/config/common"
 )
 
@@ -59,7 +58,7 @@ func TestWebhookOutput_SendEvent(t *testing.T) {
 	defer srv.Close()
 
 	u := mustParseURL(t, srv.URL)
-	cfg := config.EventRecorderOutput{
+	cfg := EventRecorderOutput{
 		Type: "webhook",
 		URL:  u,
 	}
@@ -93,7 +92,7 @@ func TestWebhookOutput_MultipleWorkers(t *testing.T) {
 	defer srv.Close()
 
 	u := mustParseURL(t, srv.URL)
-	cfg := config.EventRecorderOutput{
+	cfg := EventRecorderOutput{
 		Type:    "webhook",
 		URL:     u,
 		Workers: 8,
@@ -125,7 +124,7 @@ func TestWebhookOutput_RetryOnFailure(t *testing.T) {
 	defer srv.Close()
 
 	u := mustParseURL(t, srv.URL)
-	cfg := config.EventRecorderOutput{
+	cfg := EventRecorderOutput{
 		Type:         "webhook",
 		URL:          u,
 		MaxRetries:   3,
@@ -151,7 +150,7 @@ func TestWebhookOutput_DropsAfterMaxRetries(t *testing.T) {
 	defer srv.Close()
 
 	u := mustParseURL(t, srv.URL)
-	cfg := config.EventRecorderOutput{
+	cfg := EventRecorderOutput{
 		Type:         "webhook",
 		URL:          u,
 		MaxRetries:   2,
@@ -177,7 +176,7 @@ func TestWebhookOutput_CloseFlushesQueue(t *testing.T) {
 	defer srv.Close()
 
 	u := mustParseURL(t, srv.URL)
-	cfg := config.EventRecorderOutput{
+	cfg := EventRecorderOutput{
 		Type:    "webhook",
 		URL:     u,
 		Workers: 1,
